@@ -1,48 +1,48 @@
 {
-    "name": "budget-client",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test",
-        "test-ci": "ng test --watch=false --code-coverage"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~13.3.0",
-        "@angular/cdk": "~13.3.0",
-        "@angular/common": "~13.3.0",
-        "@angular/compiler": "~13.3.0",
-        "@angular/core": "~13.3.0",
-        "@angular/forms": "~13.3.0",
-        "@angular/localize": "~13.3.0",
-        "@angular/platform-browser": "~13.3.0",
-        "@angular/platform-browser-dynamic": "~13.3.0",
-        "@angular/router": "~13.3.0",
-        "@ng-bootstrap/ng-bootstrap": "^12.1.2",
-        "@popperjs/core": "^2.10.2",
-        "@swimlane/ngx-charts": "^20.1.0",
-        "bootstrap": "^5.1.3",
-        "rxjs": "~7.5.0",
-        "tslib": "^2.3.0",
-        "zone.js": "~0.11.4"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "~13.3.5",
-        "@angular/cli": "~13.3.5",
-        "@angular/compiler-cli": "~13.3.0",
-        "@types/d3": "^7.4.0",
-        "@types/jasmine": "~3.10.0",
-        "@types/node": "^12.11.1",
-        "jasmine-core": "~4.0.0",
-        "karma": "~6.3.0",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.1.0",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "~1.7.0",
-        "ng-packagr": "^13.0.0",
-        "typescript": "~4.6.2"
-    }
+  "name": "budget-client",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test",
+    "test-ci": "ng test --watch=false --code-coverage"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.3.0",
+    "@angular/cdk": "^14.2.7",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/localize": "^14.3.0",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "@ng-bootstrap/ng-bootstrap": "^12.1.2",
+    "@popperjs/core": "^2.10.2",
+    "@swimlane/ngx-charts": "^20.1.0",
+    "bootstrap": "^5.1.3",
+    "rxjs": "~7.5.0",
+    "tslib": "^2.3.0",
+    "zone.js": "~0.11.4"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.12",
+    "@angular/cli": "^14.2.12",
+    "@angular/compiler-cli": "^14.3.0",
+    "@types/d3": "^7.4.0",
+    "@types/jasmine": "~3.10.0",
+    "@types/node": "^12.11.1",
+    "jasmine-core": "~4.0.0",
+    "karma": "~6.3.0",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.1.0",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "ng-packagr": "^14.2.2",
+    "typescript": "~4.6.2"
+  }
 }
